(BOLT-505) pass -E to sudo when using environment variable params
This may result in tasks that do not specify input method and accept
stdin not causing a permission error when run with sudo if the user
does not have the appropriate permissions. That is preferable to running
with no environment variables set.
1. Users who don't specify input method are more likely to use
environment variables
2. Running with unset environment variables may be dangerous if the task
author was relying on parameter validation.