(BOLT-488) Submit data to Google Analytics for each command invocation
Every time the user runs a command (`bolt task run`, `bolt plan show`),
we now submit a "screenview" event to Google Analytics, where the
"screen" is the command that was run. The information included is:

a) bolt version
b) user locale
c) the "screen" run
d) a unique, random user id

IP addresses are anonymized, and we do *not* include the name of the
task or plan run.

The unique ID is stored in ~/.puppetlabs/bolt/analytics.yaml. That
file also accepts the setting `disabled: true` which will disable the
collection of analytics.

To avoid polluting the data with development and test runs, we also use
a BOLT_DISABLE_ANALYTICS environment variable, which is automatically
set when running via bundler and while running the tests.

Analytics submission is performed in the background and errors are
ignored, both so as not to disrupt normal usage of bolt. There is a
grace period of 250ms at the end of the application to allow any
outstanding event submissions to complete without introducing a
noticeable hang.

The complete data submitted is logged at debug level so that users can
be aware of the information we're collecting.