(BOLT-545) Explicitly set group based on id -g for chown
The `chown <user>:` syntax to inherit group from user is not portable.
Use `id -g <user>` to get the user's group and explicitly use it when
changing ownership to make it more portable.

Note that this will still fail on some platforms with very old versions
of `id` that don't support `-g`, such as the default `id` on Solaris 10.