(BOLT-534) Ensure sudo prompt is cancelled before continuing
If we attempt to run sudo, are provided a prompt, and don't have a
password, we abort the sudo attempt. Previously we would not cancel the
prompt, meaning any later attempt to use the same connection would be
writing to the password request instead of a shell.

This became a problem when we would attempt to chown a file (during
upload, or running a script or task) and fail when prompted for a sudo
password. As part of handling the failure we'd attempt to cleanup the
file we created, which would instead write to the password prompt and
hang indefinitely.

Now we close the channel if we can't supply a sudo password to exit the
attempt, allowing later operations to succeed.