(BOLT-533) Move facts module tests to bolt spec tests
This PR moves the spec tests from the facts module included with bolt to live with the 'general' spec tests. This is so that we can maintain some spec testing while rewriting most of the spec tests to be beaker acceptance tests for the actual puppetlabs-facts module that we're replacing the local facts module with.