(BOLT-546) Avoid mktemp to work on AIX
AIX, and possibly other such systems, doesn't have `mktemp` available by
default. That was causing remote tempdir creation to fail. Since we
already have a functioning alternative using UUIDs, we now use that in
all cases. If the user supplied an explicit tmpdir, that remains the
parent, and we default to /tmp otherwise.