(BOLT-491) Collect number of targeted nodes for each bolt command
We now include the number of targeted nodes with the screen_view hit
that we submit each time a bolt command is run. This counts the
*resolved* targets, after groups and PuppetDB queries have been taken
into consideration. This does not account for the number of nodes run as
part of a plan, unless those nodes were passed using --nodes or --query.