(BOLT-564) Collect data about calls to boltlib functions
This adds a report_function_call method to the Executor, which is now
used in each of the built-in functions included in boltlib. The
report_function_call method submits an event with category = "Plan",
action = "call_function" and label = the function name.

This will allow us to understand which built-in content is being used
most frequently and most widely.

Because Bolt uses the run_task and run_plan functions internally to run
tasks and plans from the CLI, we now pass an underscore parameter called
`_bolt_api_call` to those functions, which causes them to skip reporting
in that case. This way, we can track calls to run_task and run_plan that
happen *exclusively* from within a plan.