(BOLT-580) Bolt-transport API POC
This is a proof-of-concept for a small API that allows applications to access
SSH and WinRM, and allows them to run puppet tasks over those transports.
This manifests as a sinatra server whose lifecycle is managed by whatever
client is using the API. The server includes endpoints (currently just
`ssh/run_task`) which access the bolt executor directly to run tasks on
specified nodes and return a single result object.