(BOLT-578) Report bundled content to GA
Include the name of the task/plan being run if that name refers to a piece of built-in content. The task/plan name should not be included if it isn't a piece of bundled content, as the name may be sensitive. This data is reported to GA as an event with Category: Bundled Content, Action: Plan/Task, Label: plan/task name.