(BOLT-662) Decode base64 executables
The server example was returning an empty _output field. I initally thought this was because we weren't properly passing parameters to the executor, but it turned out it was because we weren't decoding the base64 file passed from the client. So this now decodes the file in place before scping it to the host. I also added a second task run to the example to ensure that multiple task runs with the same server work, and that not just the echo task works and returns output.