(maint) Simplify boltdir and config loading
Previously, the config class had functionality whereby defaults could be
set by passing a hash to the initialize method. This was only used for
tests, and complicated the config class by adding another level to the
"hierarchy". It also interacted in confusing ways with settings whose
defaults were already somewhat dynamic (anything depending on boltdir).

We also previously built the config object up incrementally,
instantiating it with optional default overrides, and then later
parsing a config file, and then overriding CLI options. This meant
several intermediate steps before getting to the final value of the
config.

The config object now takes a config data hash (which is the result of
loading the config file) as well as an optional hash of overrides (which
are the result of processing the CLI arguments). This allows the entire
final state of config to be computed immediately, and it also provides a
way for tests to override settings by the same mechanism that users do.

This commit also introduces a Boltdir class that captures the particular
behavior of the Boltdir, separate from the config class. The config
takes a Boltdir during initialization and uses the Boltdir to determine
some of its defaults.

We now accept only one of --configfile or --boltdir. If --boltdir is
used, then the config file will be read from <boltdir>/bolt.yaml. If
--configfile is specified, then boltdir will be set to the parent dir of
the config file. The default behavior when neither is specified remains
the same: boltdir will be found in the current directory hierarchy, or
it will fall back to ~/.puppetlabs/bolt.