(maint) Don't pass Config object around
This changes the Executor, Logger and PAL classes to accept the
config options they use as explicit arguments, rather than taking the
entire Config. That makes it easier to understand and track the flow of
config data through the system.

The Applicator class now accepts PuppetDB::Client instance rather than
that portion of the config. It uses a new method on the client class to
derive a config hash that it passes to the external compiler, which
reinstantiates the PuppetDB instance. This saves the compiler from
having to re-parse puppetdb.conf and also ensures that the client config
used by the compiler will identically match the one used during plan
compilation, regardless of what may have changed underneath. This also
makes the code above the applicator reflect the intended semantics: the
external compiler should use the same client instance used for the rest
of compilation.