(BOLT-650) Make bolt available as a service
This adds the necessary infra to deploy the pe-bolt-server as a service, including service init files, defaults, and the unit file. We also need to check if the `$HOME` environment variable exists and set paths more explicitly if it doesn't, since when the service starts that env var won't be avialable.