(BOLT-634) Fix modulepath on Windows
Windows CLI uses a different modulepath, and needs a non-Cygwin
directory to install the facts module.

Skip the apply winrm test if the controller node is a winrm target and
has its own Ruby installed, as Bolt will fail to run apply in Puppet
Agent's Ruby. That will be fixed separately.