(BOLT-684) Ensure we can operate on a ResultSet returned by apply
Previously if only an `apply` was used, nothing would refer to
`ResultSet` and the type would not be loaded. Ensure the type is always
loaded so we can use it as the result of an `apply`. We do this by
declaring the return type of `run_plan`; running a plan is when we would
care about the data type returned by `apply`, and `PlanResult`
references `ResultSet` ensuring it's loaded.