(BOLT-635) Use separate thread pool for compiles
Use a separate thread pool when compiling so we can limit it to a lower
rate than execution (which is not CPU bound). Defaults to the number of
threads on the system.

Additionally, errors during compilation will now be rolled into a
`ResultSet`, rather than causing `apply` to erroneously return a single
`Result`.