(maint) Delay PuppetDB client config validation
Previously, we used an OnAccess object to defer loading and creation of
the PuppetDB client and its config. That was done lazily because loading
the config object caused validation automatically, meaning trying to
create a PuppetDB client without an existing would fail, even if the
user didn't want to query PuppetDB in the first place.

That behavior also made it impossible to derive the config back from the
client to pass along to the external compiler when using the `apply`
keyword; if the config was invalid, it would fail as soon as we tried to
retrieve it, again regardless of whether the user actually needed to
query PuppetDB.

This commit changes the interaction between the config class and the
client class, such that the client simply stores a reference to its
config and fetches values when needed (only when actually trying to make
a query). That allows us to safely instantiate and pass around an
instance of the client, as well as recover its config to pass to the
external compiler, even in the case where the config is invalid.