(BOLT-698) Fix powershell output ordering by using a synchronous exec
Run executables on the target over WinRM by executing them directly in
powershell. Prior attempts to capture stdout and stderr asynchronously
ran into a problem that powershell event ordering is not guaranteed. We
lose the ability to specify a timeout, but we already don't have that
for invoking a powershell script.

Adds tests to ensure many lines are output in the right order.

Set output encoding so UTF-8 tests work as described in
https://blogs.msdn.microsoft.com/powershell/2006/12/11/outputencoding-to-the-rescue/.