(BOLT-523) Create default text domain when initializing Puppet
When Puppet is first loaded, it creates a gettext domain called
"default-text-domain". When clearing environment-specific domains
between tests, it will make sure to preserve that domain. It has no
such deference for any other domains, however, including the
"master_domain" that is created when loading r10k.

Because Puppet only creates this default domain when *loaded*, it may or
may not wind up being the actual default domain when executing. When we
clear domains between tests, if the default domain is "master_domain"
(because we required r10k after puppet), then it will be deleted and
gettext errors will occur.

To avert this load order dependency, we now explicitly create Puppet's
default domain when initializing Puppet via PAL.