(BOLT-691) Fix running task with tty echoing input
When running a task with tty and using sudo to execute as a non-root
user, if stdin isn't consumed (like in the `facts` task) but is sent
over SSH it appears to be echoed back. Disable echo mode when using tty
to ensure it's not echoed.

Note that when using `--tty`, Puppet prints errors on stdout instead of
stderr.