(BOLT-698) Force encoding on result to match external encoding
When running Bolt from Windows on Windows with Ruby 2.3, the input gets
transformed from UTF-8 to the target encoding but does undergo the
reverse when it comes back. So in AppVeyor we end up with Windows-1252
encoded characters in a string that Ruby thinks is UTF-8 and get an
exception. Force the encoding to the Windows external encoding then
convert to UTF-8 before comparing.

This issue appears to be fixed with Ruby 2.4.