(BOLT-640) Build and sync plugin tarball to agents
When using the apply() keyword, we now copy the contents of the `lib`
and `files` directories of every module to the agent. This allows native
types and providers to be used, as well as for puppet:// file sources to
resolve correctly.

This is accomplished by packing those files into a gzipped tarball,
which is then passed as a parameter to the apply_catalog internal
task. That tarball is then written to disk on the agent, unpacked, and
set as the modulepath for the catalog application.

Continue to mock Puppet.lookup for unit tests, providing an empty list
when querying available modules.