(BOLT-640) Fix filepath setup to be usable on many platforms
In a previous commit filepath was updated to generate a unique name each
test run. However in the SSH tests it was created based on the Bolt
controller's concept of tmp, which may not be simply portable. Generate
a unique directory name in a known location that should be safe for SSH
targets. Do the same with WinRM for consistency.