(BOLT-641) Use consistent HOME dir when testing Bolt
Previously we used two different HOME locations when running Bolt on
Windows. This was to provide path to modules that we could put on the
command-line that both Cygwin and Bolt would understand. However with
the move to using Bolt to install from a Puppetfile to Boltdir, we need
to use a single location for Boltdir. Update test helpers to use the
Windows paths to the default Boltdir.

Also, always ensure the default_boltdir exists before we try to copy to
it.