(BOLT-757) Include custom facts and features in apply
Ensure external facts are copied to targets during apply, that the apply
sets up Facter to find them, and that the lib dirs of modules are
available when resolving `puppet/feature` with the autoloader and
requires statements from providers.

Also make it easier to debug `apply_catalog` by accepting a file name on
the CLI as an alternative to passing args via stdin.