(BOLT-723) Add winrm run_task endpoint
This adds a new `/winrm/run_task` endpoint to pe-bolt-server which runs a task on a node over the winrm transport. It currently writes a `.ps1` file to the bolt controller system, and uses that as an executable to run the task over winrm. It also sets the `input_method` to 'powershell' in order for the parameters key to work.