(BOLT-723) Document changes to the bolt-server API
This commit documents 2 changes to the bolt-server API:
* The addition of a `/winrm/run_task` endpoint, which is very similar to the `/ssh/run_task` endpoint but whose target object has different options
* The addition of a `file` JSON object in place of `file_contents` in the task object.

The latter change is because we need to know the filetype to run tasks on windows hosts. Long term we want to get this data directly from puppetserver, but aren't sure we can make this change in the time frame we have. For now we still need the file_content, in addition to the filetype, and will accept other keys provided by the [puppetserver task api](https://github.com/puppetlabs/puppetserver/blob/master/documentation/puppet-api/v3/task_detail.markdown).