(maint) Hard-code GEM_PATH for bolt-server
GEM_HOME appears to be appended to GEM_PATH. If a different version of
Bolt is installed in puppet-agent, that impacts how bolt-server
functions. Set GEM_PATH instead to ensure we use the bolt-server
specific gems first.

This ties bolt-server's service config more closely to the puppet-agent
Ruby setup, since we have to include its GEM_PATH as well.