(BOLT-812) Bolt correctly identifies missing Puppet on newer Powershell
On newer versions of powershell, such as the one in Nano Server that we
use for local testing, the error reported when trying to run 'apply'
changed from `Could not find executable 'ruby.exe'` to `The term
'ruby.exe' is not recognized as the name of a cmdlet`. This does not
seem to apply to older versions of Powershell such as used in AppVeyor
CI. Correctly match the new error string and report that Puppet is not
installed when trying to use `apply`.