(maint) Refactor winrm transport
This commit refactors the winrm transport to mirror the ssh transport. Specifically the way files are uploaded to the target to support running tasks with the origin of a bolt server request instead of a file on disk.

Additional functionality is being added to the transports to support bolt-server work. Harmonizing the general layout of the winrm and ssh transports removes some friction while adding new functionality intended to be supported by both transports.