(BOLT-750) Remove apply module and apply::resource task
This module was an experimental inclusion to provide basic functionality
to reuse existing Puppet resource types. Since we've now introduced the
apply() keyword as a general, supported solution to this problem, the
apply module is no longer necessary. We're choosing to immediately
remove this rather than add a deprecation based on the fact that we have
no known users of the feature.