(TBD) Provides a '--stream' option for streaming output
Adds a `stream` transport option for streaming output that's implemented
on SSH and WinRM. Augments logging to include target name, and when
`stream` is enabled outputs stdout and stderr from executing commands to
a special `stream` log level (that can be used to direct it to a
specific file).

Fixes #102.