(BOLT-719) Allow modulepath relative to Boltdir
Previously, relative paths in the bolt.yaml file were treated as
relative to pwd, which is almost never what's expected. We now treat
them as relative to Boltdir, allowing a modulepath like site:modules to
be used to represent .../Boltdir/site:.../Boltdir/modules, regardless of
where the Boltdir is located on disk.

When modulepath is specified on the CLI, relative paths are still
interpreted as relative to the pwd, which should be less surprising for
new users than making it implicitly based on a directory (Boltdir) they
may not even know exists.