(TBD) Bolt Service Dockerfile
This wraps bolt-server in a container.  It uses a simple shell script to
install and configure defaults.

Long term this should be built in the same way as rpm,dpkg packages are
built to avoid issues such as systemctl not running in most containers.

To build this requires docker and golang

    ```
    cd bolt-service
    make
    ```

To run

    ```
    make run
    ```

See the README.md for more details