(BOLT-766) Add 'site' to default modulepath
Previously, the default modulepath was just `modules`. That made it
difficult to use a Puppetfile with `bolt puppetfile install` while also
using some custom module content, as every custom module needed to be
added to Puppetfile with `local: true` simply in order to be preserved
when updating.

Now, we use `modules:site` as the default modulepath, which provides
users a location to install modules via the Puppetfile (`modules`) as
well as a location for custom content that should not be managed with
the Puppetfile (`site`). `bolt puppetfile install` will always install
to the first element of the modulepath, so we break the typical Puppet
convention and put `modules` *first*.