(BOLT-656) Add Executor mode without a thread pool
We use the executor to run tasks from bolt-server. In that scenario, the
web server is handling concurrency and we want to avoid extra threading
that won't be used.

Executor as used in Bolt requires a positive concurrency. Add a mode
that disables threading when concurrency is 0. In this case any run
requests will iterate sequentially over targets.