(BOLT-656) Use a single Executor without a thread pool in bolt-server
Create an Executor once on startup for bolt-server and use it for all
requests. The Executor is thread-safe as long as you avoid `start_plan`,
`finish_plan`, `without_default_logging`, and changing `run_as`.

This change greatly reduces memory churn for requests, and results in a
stable number of threads throughout the server lifetime.