(BOLT-840) Don't sync built-in plugins for apply()
Because we vendor the externalized "core" Puppet modules, we need to
avoid syncing them to agents. These plugins are included out-of-the-box
in puppet-agent 5 (where they are core) and 6 (as modules). That means
it's unnecessary to copy any of this content to agents. Moreover,
because the vendored modules are only intended to be used by Puppet 6
agents, they can cause problems when syncing to Puppet 5 agents, where
they can supersede the built-in plugins with potentially
backward-incompatible code.

We therefore skip our own vendored module directories (bolt-modules/ and
modules/) when copying plugins. This has the added benefit of slightly
shrinking the plugin tarball we ship to agents.