(BOLT-840) Include "vendored" Puppet types
In Puppet 6, many types and providers have been removed from core into
their own modules. Those modules are included in the puppet-agent
package so the out-of-the-box experience remains the same. In order to
preserve that same experience when using the apply() functionality, we
now vendor the same modules in Bolt.