(maint) Fix transient failures in executor spec test
We are seeing transient issues around the default sleep time for checking threads used by executor. When we first observed the transient failures we bumped the wait time from 0.1 to 0.2 seconds. Instead of bumping the wait time again Nick Lewis suggested looping until the threads are done with a max timeout to avoid continuing to hard code wait times that may not be needed in tests.