(BOLT-852) Ensure user and password aren't mangled
When a user or password are passed via the URI, Bolt unencodes them to
allow passing complex strings in the URI. It shouldn't do that for
users and passwords passed via CLI flags or inventary, but did, meaning
passwords that include things like `%F3` get mangled and don't work. Fix
so only URI components are unencoded.