(BOLT-857) Properly override environment in all cases
Previously, when using the apply() function, it was possible for Puppet
"features" to be loaded from the default modulepath, rather than from
the temporary modulepath we set up. This could cause incorrect versions
of features to be loaded or for necessary features to be missed.

This happened because there are two methods that parts of Puppet use to
find the current environment:
* Puppet.lookup(:current_environment)
* Puppet.lookup(:environments).get(name)

We override the production environment with our custom modulepath and
set that as :current_environment in the Puppet context. However, that
doesn't affect the :environments value in the context, which still
contains the default directory environments. Thus, any code that looks
up the environment that way will find the original version of the
production environment with its default modulepath. This includes the
code that loads features.

We now explicitly override :environments with a static environments list
that only includes our modified production environment.