(BOLT-643) Use module-qualified name for apply tasks
Because these tasks are actually mocked in Bolt, their name is
irrelevant. However, in order to run them in PE, a task with the same
name needs to actually be installed. These tasks have been copied into
the puppetlabs-apply_helpers module for installation in PE.  Therefore,
we need to refer to them by the names `apply_helpers::custom_facts` and
`apply_helpers::apply_catalog` in order to work in PE.