(BOLT-816) Enable control repos as Boltdirs
For existing Puppet users, the control repo is a source of Puppet
content including modules and hiera data. It's helpful to be able to
use Bolt from within that context and pick up all of the content that's
deployed in the Puppet environment. However, doing that currently
requires creating a Boltdir with a bolt.yaml that refers to the hiera
config and modulepath using relative paths like ../modules.

We now look for a bolt.yaml in the current directory (if there is no
Boltdir) and treat the local directory as the Boltdir if one is found.
This allows the defaults for hiera config and modulepath to work when
running Bolt from a standard control repo.