(BOLT-914) Clean up Bolt-as-a-library includes
When loading Bolt as a library, the main entry point is `bolt/executor`.
It previously required Config and CLI, both of which are unnecessary for
the Executor. Clean up config so `bolt` and `bolt/executor` don't bring
in unused files, especially the CLI (which loads r10k and is not as
careful about side-effects).