(BOLT-858) Add `bolt apply` to apply a manifest directly to targets
Previously, users could write a plan that used the `apply_prep` and
`apply` functions to apply a snippet of manifest code to targets. This
is a convenient and common workflow for both applying simple snippets of
manifest code to targets as well as for testing out local changes to
code. The overhead of wrapping the manifest code in a one-off plan with
the necessary ceremony to accept parameters, etc is onerous for that
basic use case.

This introduces a `bolt apply <manifest>` command that can be used to
take a snippet of ordinary puppet code, not wrapped in a plan, and apply
it remotely to targets. This allows to a smooth workflow for making
local changes to Puppet code, testing them on nodes through Bolt, and
then pushing them up to a central control repo.