(BOLT-858) Support `bolt apply -e` with a snippet of code
This allows users to easily run simple code like `include <classname>`
without having to write it to a file. We copy the `puppet apply` option
name of --execute/-e for familiarity and consistency.

This fails if a user specifies both a snippet of manifest code as well
as a manifest file.