[FM-7504] - Put a conditional around Winrm version
Due to a new version of Winrm being released that has a higher Ruby dependency (>=2.2.0)
installations of Bolt are failing on Puppet 4, considering that Puppet 4 ships with Ruby 1.9.0.

This is an attempt to fix this issue by pinning to the last version of winrm that will work with older Ruby,
while allowing newer versions to install the latest 2.x release.