(BOLT-585) Use parser directly to parse Puppet code
Previously, we were creating a PAL catalog compiler, which does a lot of
upfront work that is unnecessary for the task of simply parsing a string
of code. This adds a method to Bolt::PAL that constructs a Puppet parser
object directly and uses it, which cuts the parsing time from ~3 seconds
down to ~0 seconds.