(BOLT-925) Scrub stack trace from bolt server errors
Remove stack traces from bolt-server responses by removing them from the result status hash in bolt server. This should preserve the trace in log messages. If a stack trace is included in the result of a task output (for example a task that throws and exception durring it's execution) it will still be included in the response.