(BOLT-942) Allow target to be added to group during plan execution
This commit adds an `add_to_group` method callable from Target objects in plans. The method can be called from a `Target` object with the argument of the group to add the `Target` to. Similarly the method can be called with two arguments the first being the string representation of a target and the second the group to add to. When a group does not exist the `Target` will not be added to ANY group. If a `Target` of the same name exists in the requested group it will be overridden by the updated `Target`.