(maint) Remove redundant testing
These tests previously were implemented to test both that Puppet would
correctly load metadata or produce an error for invalid metadata, and
that the transports would correctly copy metadata. However the transport
shared tests already include a test that metadata is copied with the
expected layout. Remove redundant tests and update the shared transport
tests to be more robust.