(BOLT-791) Support remote tasks.
This adds a remote transport to support revision 4 of the task spec. The
remote transport will proxy a task to the node defined as run-on its
config. Support for remote tasks is implemented as a transport to make
it easy to add remote targets within the existing inventory file
format. This required some refactor of the config and transport classes.

The resource_api gem is part of the puppet package and needed to load
some device modules such as panos that will contain remote tasks.