(BOLT-510) Add node aliases to inventory
An `alias` property can be set on a node in inventory as an alternate
name whereby you may refer to that node. `alias` may be a string or an
array of strings; each string is restricted to alphanumeric characters
identical to group names. Any of the listed aliases may refer to that
node. The alias can be referred to when looking up a node via
`get_targets`.

It's an error for an alias to match the name of a group or node name, or
for an alias to refer to multiple different nodes. Node names continue
to be unique, i.e. multiple node entries that use the same name refer to
the same node and aggregate configuration among them, including aliases.