(docs) Recommend `gem install --user-install`
Our previous instructions for using gems with Bolt required
root/Admnistrator access on the host. That is not required to install
Puppet modules, which meant you might be able to install a module but
not the gems required for it to work. Provide instructions that work as
a non-privileged user.