(BOLT-608) Refactor `default_input_method` for powershell
When running over the WinRM or Local transport the default for a powershell executable should be the powershell input method. This commit refactors the 'default_input_method' in the transport base class as well as where it is overridden in the WinRM and Local transport classes to accept the executable path in order to determine file type.