Helper file location details
PR-793 got a little confusing because it was merged at the same time as remote tasks, which was inserted -- I think -- in the wrong place on this page. Please confirm the location of this added helper file detail, as well as the formatting change and edit. 

This is already in docs, if you'd like to confirm there too: https://puppet.com/docs/bolt/1.x/writing_tasks.html