(BOLT-817) Support array of directories for modulepath
This allows the `modulepath` setting in `bolt.yaml` to be specified as an
array of module directories in addition to the previous behavior of a
PATH-like string. Because the PATH separator varies by platform, this
allows users to write a single `bolt.yaml` that will work on both *nix
and Windows systems.