(maint) Ensure targets have puppet-agent feature before service task in apply_prep
The most reliable task implementation for managing puppet with the service task is the ruby implementation which requires the puppet-agent feature. This commit sets the `puppet-agent` feature on all targets that 1) did not previously have the `puppet-agent` feature set, 2) have been determined to have an existing puppet agent or 3) have just had an agent installed.