(BOLT-1049) Update acceptance tests to service::init task
The most reliable way to query the status of puppet using the service task is the ruby implementation. This commit uses updated service, package, and puppet_agent modules for git based testing and updates the apply tests to use the service::init task by wrapping in a plan that sets the puppet-agent feature on target nodes.