(maint) Update puppetdb container setup
The latest and greatest puppetdb container requires some changes to how it is set up as it uses a stripped down alpine base image and builds puppetdb from source instead of a package. This commit changes the entrypoint to run only the ssl-setup script and then invoke the closure service.