(BOLT-153) Support SMB file transfer
WinRM sucks for file transfer because it base64 encodes the payload,
compresses it as a zip file, transfers the contents via Powershell in
SOAP/XML messages, and decompresses on the remote side.

This commit allows bolt to upload files using SMB using the Rapid7 gem.
If anyone understands SMB, it's Rapid7/metasploit. The WinRM connection
lazily generates an SMBClient if a file is uploaded. The connection
will only login once, but connect to and disconnect from the share for
each uploaded file.

Before this commit transferring a 75MB jre takes 58.7 seconds to a local
Windows VM. With this commit it takes 6.2 seconds.

By default, WinRM connections continue to use WinRM for file transfers.
To enable SMB, add the following to the inventory file:

    winrm:
      file-protocol: smb

When uploading a file, the destination path should be specified as
either a drive letter or UNC path. If a drive letter is given, then bolt
will use the administrative share, usually `\\<host>\C$`. Note this
requires administrative privileges. To upload a file as a non-admin
user, enter the share as a UNC path. If bolt is running on a Windows
host:

    PS C:\> bolt file upload C:\path\to\file \\host\share\path\to\filebolt file upload C:\path\to\file \\host\share\path\to\file

If bolt is running on a non-Windows host, backslashes in the destination
path need to be escaped:

    $ bolt file upload /path/to/file \\\\host\\share\\path\\to\\file

Or forward slashes can be used (for drive letter or UNC paths):

    $ bolt file upload /path/to/file C:/path/to/file

- Implements SMB based tests for file uploads.

- Implement test confirming SMB + SSL is currently unsupported

NOTE: SMB port is currently not configurable, which doesn't allow tests
      against local Vagrant to pass.

      Recursive directory copy is not implemented with SMB support (yet)