(BOLT-1107) Use internal orchestrator API for running plan tasks
Previously, the plan executor service used the public orchestrator API
to run plan tasks and submit plan results. This required it to follow
the same rules as a user running plans; specifically, it needed to
authenticated with an RBAC token and was only allowed to operate on
plans that were started by the same user for whom its RBAC token was
issued.

Now the service uses the internal API with certificate authentication to
connect as the API user and operate on plans with complete authority.
Because the public orchestrator-client library doesn't support features
we *need* like certificate authentication and the internal command API,
and it *does* have some automatic behavior around config loading that
we *don't* want, we now use a JSONClient (from the httpclient library)
directly to talk to the orchestrator service from the plan executor
instead.

We now create the http client during app startup time and pass it
through when creating an executor / orch client. That simplifies mocking
the http client in tests and also ensures that if we were to run
multiple plans at once, they would share a single http client.