(BOLT-1118) Mark internal modules in show-modules output
This commit updates the `bolt puppetfile show-modules ` command to mark modules
that are packaged with Bolt as opposed to installed by users. This is done by
displaying the group that the module belongs to instead of the path in human
readable output. JSON output contains an additional internal_module_group key
that holds the name of the group.