(BOLT-1057) Standardize BoltSpec::Run helper method signatures
Previously, these methods had confusing and inconsistent signatures
using a mix of optional parameters and keyword arguments. In some cases,
the optional parameters were themselves expects to be hashes, which
conflicted with the keyword arguments. That caused confusing behavior
when invoking the functions with only some arguments, as a hash intended
to be a positional argument could be interpreted as an _invalid_ keyword
arguments hash, or vice versa.

These functions now always require their `params` (or equivalent)
argument, leading to slightly more verbose but always clear
invocations. The functions which previously expected an options hash in
addition to their keyword arguments now expect the options hash to
be provided as a keyword argument named `options`.