(WIP)(BOLT-1129) Generate metadata.json files for internal modules
This commit adds a metadata.erb template to the internal bolt modules that
can be rendered to produce metadata.json. The Rakefile has been updated with
a rule that executes the rendering when one of the following is true for a
module:

  - The metadata.json output is not present.

  - The metadata.erb template has been updated.

  - The lib/bolt/version.rb file has been updated.

The generated files have been added as dependencies to the standard build task
such that `rake build` will create or update the metadata.json files before
creating a gem package.