(BOLT-1108) Add site-modules to modulepath
Previously site specific modules were expected to be in the default `site` directory. This was just a convention and the name `site` doesn't clearly convey what it is for. This commit adds a new default `site-modules` directory to the modulepath. This makes it more clear to users that this is a directory that modules go in.